Implementar módulo de préstamos: crear controlador, vistas y migración para la gestión de préstamos de libros.

+++ database/migrations/2026_02_18_072129_crate_table_libros.php
@@ -4,8 +4,6 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-use function Laravel\Prompts\table;
-
 return new class extends Migration
 {
     /**
@@ -13,22 +11,26 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('categorias', function (Blueprint $table) {
-            $table->id();
-            $table->string('nombre', 100);
-            $table->timestamps();
-        });
+        if (! Schema::hasTable('categorias')) {
+            Schema::create('categorias', function (Blueprint $table) {
+                $table->id();
+                $table->string('nombre', 100);
+                $table->timestamps();
+            });
+        }
 
-        Schema::create('libros', function (Blueprint $table) {
-            $table->id();
-            $table->string('titulo', 225);
-            $table->string('autor', 100);
-            $table->string('isbn')->unique();
-            $table->string('editorial', 255);
-            $table->smallInteger('estatus')->default(0);
-            $table->timestamps();
-            $table->foreignId('id_categoria')->constrained('categorias')->onDelete('cascade');
-        });
+        if (! Schema::hasTable('libros')) {
+            Schema::create('libros', function (Blueprint $table) {
+                $table->id();
+                $table->string('titulo', 225);
+                $table->string('autor', 100);
+                $table->string('isbn')->unique();
+                $table->string('editorial', 255);
+                $table->smallInteger('estatus')->default(0);
+                $table->timestamps();
+                $table->foreignId('id_categoria')->constrained('categorias')->onDelete('cascade');
+            });
+        }
     }
 
     /**
